FalxApi updates and fixes (#13)
* Fixed a bug in FalxEventStore caused by a previous cleanup commit.
Updated this function in FalxEventStore so that it always return an empty array instead of null
when there are no events stored:
public List<AggregatedFalxMonitorEvent> aggregatedEvents(String eventName, boolean allowPartialDays)

Bumped version number and code.
versionCode 4
versionName "0.1.2"

* Tested a JobScheduler that periodically looks up aggregated events from the Falx lib
and processes them.

Updated some logging function to not log a lot of information which was previously used for debugging.

* Moved Realm.init from FalxApi to FalxEventStore.

In the test app added Robolectric to develop a Shadow for FalxApi that other apps
can use to mock out FalxApi calls with the shadow if needed.

* Refactored code so that we can mock FalxEventStore for unit tests.
Added Module inside the UtilComponent to inject FalxEventStorable to classes.

FalxApi is updated to not not add duplicate Monitors.
Added unit tests for FalxApi.addMonitors(). (FalxApiTest.java).

Fixed SessionTrackingTests.java tests that were failing.