LIFE-15064: Crash in Falx (#28)
There was a OOM issue caused by Realm. The cause was EventArguments was not getting deleted
when we cleaned up FalxEventEntity objects because Realm does not support cascading deletes if
there are lists of Realm objects inside it. Fixed the deletion function, and added a function to clean the database.

+ Updated Realm to v4.2.0